requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Async
asyncio>=3.4.3
aiohttp>=3.9.0
httpx>=0.25.0

# Database
sqlalchemy>=2.0.23
asyncpg>=0.29.0
alembic>=1.12.0

# Redis
redis>=5.0.0
aioredis>=2.0.0

# LLM
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0
tiktoken>=0.5.0

# RAG
chromadb>=0.4.0
langchain>=0.0.340
langchain-community>=0.0.10
sentence-transformers>=2.2.0
pypdf>=3.17.0
python-docx>=1.1.0
unstructured>=0.11.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Task Queue
celery>=5.3.0
flower>=2.0.0

# Monitoring
prometheus-client>=0.19.0
structlog>=23.2.0
sentry-sdk>=1.38.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Utils
python-dotenv>=1.0.0
tenacity>=8.2.0
orjson>=3.9.0
